Version 1.1.4 Release Notes:
Add feature so that deleted flights can be restored to the flight sheet (deleted flights are only retained in memory and will be lost if the program is restarted).
As part of the "restored deleted flights" feature, flight numbers are never reused. That is, deleted flights leave a numbering gap in the flightsheet that can only be filled by restoring the deleted flight. Note that the printed flight sheet has consecutively numbered flights that are always in take-off time order (see below).
Fix bug where printed flight sheet was not always in take-off time order. Program now automatically sorts by take-off time before printing.
Show "non/club" in estimated cost column when the pilot(s) are not club members.